requirements.txt
pymupdf
python-dotenv
kokoro 
soundfile
groq
huggingface_hub
pydub
numpy
# --- IGNORE ---